Testing: Use PORT2 instead of PORT5 for Solr during tests:
The docker service layers that were added (msgconvert, weasyprint, etc..)
use ports 5-8 for their docker containers. Because Solr already uses PORT5
as well, this leads to occasional port conflicts on Jenkins.

Because the ports for the service layers are a nice sequence, we change the
Solr port instead (to PORT2, which is currently unused).