Testing: Make sure layer names don't start with the name of another layer:
The --layer flag of collective.xmltestreport / zope.testrunner greedily matches
layer names by default.

For this reason, when a layer name starts with a prefix that is itself the name
of another layer, we run into issues with bin/mtest because tests on that
layer will be run multiple times, may not be excluded as expected or may run
in parallel even though they shouldn't.