Sequential tasks: Only allow subtasks to be started when parent is in progress:
This prevents subtasks in a subprocess to be started manually by the user
unless the containing parent task is already in progress. This avoids getting
the task structure into a state where the logic to automatically start the
next task fails. That logic assumes that the next task to be started is in a
'planned' state, and trips up if it has already been started manually.