Add `@@oc_checkout` view that redirects to oc: checkout URL:
This view allows 3rd party applications to produce links that should directly
trigger a checkout using OfficeConnector on a GEVER document.

In case a necessary precondition is not satisfied (e.g., document is already
checked out by someone else), the view just redirects to the regular document
view, which should give the user some indication as to why the checkout didn't
work.