Untangle address block DocProperties from OgdsUserToContactAdapter:
Until now, the address block for OGDS users was generated from an
OgdsUserToContactAdapter, because that was the only object that appeared in
'recipients_data' or 'sender_data' that had a reference to an actual OGDS
user.

We now add an actual OGDS user to that data, give it an IDocPropertyProvider,
and produce the address block doc properties directly from that.

This will make it easier to remove the SQL contacts functionality.