Fix backported test:
One of the tests that were backported for 4teamwork/opengever.core#7624
depends on behavior that isn't on the 2022.15 branch yet:

It assumed that group memberships for deleted groups get removed during OGDS
sync, but that fix was only implemented recently (4teamwork/opengever.core#7573)
and isn't present on the 2022.15 branch yet.

The test therefore needs to be adapted accordingly.