ogcore docker image: Send ftw.upgrade's stats log to /dev/null:
We symlink ftw.upgrade's upgrade_stats.csv to /dev/null, because otherwise
ftw.upgrade will try to create it during upgrades, and will fail because of
insufficient permissions.

Because we don't actually care about the contents of this logfile, we can just
discard its content instead of redirecting it somewhere else.