Cast datetime values to date objects for customproperty date fields.
Setting date values to a custom property is not possible when using the resolver_custom_after_transition_hook of the IDossierResolveProperties because it is not allowed to use pythons datetime.date() function. We can only use DateTime.utcdatetime(). Therefore, we need to cast the datetime value to a date if the underlying schema field is of type Date.