Handle review_state mismatch between successor/predecessor pairs gracefully
In some caes there is a mismatch between successor/predecessor pairs,
which leads to a WorkflowException when trying to sync the workflow
transition to the remote site, via remote_request. This cases needs to
be fixed manually via debug mode.

To prevent this support effort, we fix this review_state mismatch
automatically.