Implement IGroupIntrospection for OGDS auth plugin
IGroupIntrospection is needed to allow setting permissions for OGDS
groups in the sharing view. Also makes OGDS groups show up in the Plone
control panel as well.

The methods aren't cached currently, as we expect they are used rarley.
Caching could be easily done though.

getGroups() is implemented in a naive way and could be optimized to not
query each group separately, but we are not expecting to use it.