Schema dumps / bundle import: Introduce OPTIONAL_ROOT_TYPES:
Currently, opengever.inbox.inbox is the only OPTIONAL_ROOT_TYPE:
These *may* be constructed directly at the site root, but don't always have to.

(an inbox may be constructed at the Plone site root, or inside an inbox
container)

They are slightly different from PARENTABLE_TYPES, which can *always* be
unambiguously parented to a container that is guaranteed to exist.