Bundle import: Map principal names to ids in local roles:
For principal names that appear in local roles (the "_permissions" property
in bundle items), map their names to ids based on OGDS users/groups.

This is required to support new-style deployments with OGDS auth plugin and
user/group names and ids that are distinct from each other (as opposed to
legacy style deployments where they are identical).