Fix tests that fail because of changed group memberships:
In our integration test fixtures, we create OGDS groups projekt_a and
projekt_b, mainly to use them as groups for teams. But we also add OGDS users
to them.

Previously, when OGDS groups were not use in PAS for group enumeration, those
groups didn't really have an effect, because they were not in source_groups.
But now that the OGDS auth plugin is installed, those group memberships get
recognized, and the users are actually considered to be in those groups too
by PAS, so their group memberships changed with the introduction of the OGDS
auth plugin in tests.