Fix docproperty writing when creating a doc from template.
By avoiding the creation of the initial version before writing the
docproperty values, we fix an issue when creating a document from
template. Currently when a document gets created from
template, the docproperty values are not written to the initial
version of a document, but only in the working copy. This leads to a
wrong document, when closing the document without doing any change
after generation. Because with the recently changed office connector
behavior, this cancels the checkout and drops the changes in the
working copy.