Fix readonly file download tests:
It seems that the presence of the OGDS auth plugin causes an additional
write on login in some cases (probably a time stamp like last login time).

We therefore commit the transaction as part of the test arrangement, like we
already do in another test, to get any pending writes out of the way before
the actual test.