Fixture: Set group title when creating committee OGDS groups:
Instead of the builder default 'Test Group', the committee OGDS groups now
get proper titles.

Because the title affects ordering of some vocabularies, for some of the tests
that only assert on the first batch of results it might look like the groups
changed entirely from looking at the diff - but they didn't.