Create an OGDS user for TEST_USER_ID created by p.a.testing:
Once we disable user enumeration for source_users in tests, the standard
user TEST_USER_ID created by plone.app.testing will not be enumerated any more.

Because way too many tests (and testing infrastructure code) depend on this
user being present, we have to create it as an OGDS user, so that it will be
enumerated by the OGDS auth plugin.

This however will lead to a few tests failing, because the default in the
'ogds_user' SQL builder is to use TEST_USER_ID if no explicit userid is given.

This will cause any test to fail that uses the 'ogds_user' builder without
an explicit userid, since that user now already exists as part of the fixture,
and trying tot create it again will lead to an SQL integrity error.