Fix tests that attempt to create an OGDS user with the default userid:
Because we now have an OGDS user with the userid TEST_USER_ID in the fixture,
and the 'ogds_user' SQL builder uses TEST_USER_ID as the default, tests that
create an OGDS user without specifying an explicit userid are failing.

(Changing the default for the 'ogds_user' SQL builder to something else leads
to 100+ test failures for various reasons).