Fixture: Drop the Plone User for 'inactive.user':
The concept of 'inactive' is only applicable for OGDS users. But when we create
both a Plone user and an OGDS user that is inactive with the same ID, this
causes problems:

For example, when the source_users plugin is asked to list all its users (say,
by the `@users` endpoint), then it will find 'inactive.user' in its IDs. But
because actual enumeration is handled by a different plugin (OGDS auth), and
that one filters out any inactive users, source_users.listUsers() will return
None for that user, and make some PAS APIs fail.