Nested task templates: Also pass docs to subtasks that get started:
This will make sure that if a sequential task is started because it is the
next one in line, and the user selected `pass_documents_to_next_task`, the
documents will be passed (i.e. set as relatedItems) not just to that next
task, but also to any subtasks that get started if the next task happens to
be a task template.

So in short:
If the user selects `pass_documents_to_next_task`, the documents will always
be passed to exactly the same tasks that automatically get started.