Also update modification date when updating changed timestamp:
The solr-sync command uses the modification date indexed in the catalog to
check for items that are out of sync. However, the modification date isn't
updated as consistently as we thought. For example, triggering a workflow
transition will not update the modification date on its own.

We therefore also update the modification date any time the `changed` date is
updated, which should catch many of the user-visible changes to content.