Create site and run upgrades on startup
Creates a new policyless deployment if no existing Plone site is found
and runs pending upgrade steps on startup.
For now it's only enabled for Docker containers but it can be enabled
for Buildout deployments as well by setting the environment variable
`OG_AUTOSETUP=1`.

To ensure only one instance performs installation or upgrades in a multi
ZEO client setup, a lock marker is written to the database. All
instances not having the lock, wait until the lock is released before
getting ready to serve requests. The lock is checked regularly with a
maximum intervall of 5 minutes.

On the first start with multiple instances and an empty database, a
ConflictError will probably occur. This is because Zope itself does
setup it's application object in the database but doesn't handle
conflict resolution. Instances running into a ConflictError need to be
restarted. Or better, use only one replica on initial setup:

```
docker compose up -d --scale ogcore=1
```