Require dossier transfer token to match token of a particular transfer:
When validating a dossier transfer token, we now also verify that the token
that is being validated matches the token stored on the particular transfer
in question.

This was sort of validated before as a side effect, because we verified that
the transfer ID from the signed claims matched the transfer's ID, but now
we also explicitly require the token contents to match.

In addition, the TokenManager will now update the persisted token on a
transfer when issuing a token for it.