Fix OGDS listing date filters to also work on Oracle:
We used to pass an ISO formatted date string to SQLAlchemy
when constructing the BETWEEN clause. Apparently this datetime
string format is not appropriate for Oracle, and it complains
with

  ORA-01861: literal does not match format string

Parsing the date strings submitted by the frontend into
Python dates and passing those instead lets SQLAlchemy format
them correctly for Oracle.