Disposition: Include stats about document count and cumulative file size:
These stats are persisted on the disposition, and tracked by dossier. They get
updated when dossiers are added or removed from the disposition. Since the
dossiers are in a locked state 'offered' once added to a disposition, these
stats should not get out of date.

Stats are queried from Solr, with one query per dossier.