Remove Rubygems from testserver buildout
We don't install Sablon with Rubygems in opengever.core for a while.
But in gever-ui, the testserver is still installed this way.
With this change we can get rid off Rubygems there as well.

The testserver now automatically starts a Sablon container if no
SABLON_URL was given through the environment.