Fix Solr tests
Required due to https://github.com/4teamwork/opengever.core/pull/7958
Didn't get noticed because no new ogsolr image was built, which is used
in tests.

The regular `@solrsearch` endpoint is mainly affected by the changes,
but it's no longer used in the UI, so we just remove the assertions
as fixing it would reintroduce the performance problems.
Currently we are only using the `solrlivesearch` endpoint.