Fix OGDS sync for multi-tenant setups:
OGDS sync was broken for multi-tenant setups, because dispatching the remote
request to update the sync timestamp failed when attempting to create an
authentication token for a userid of `None`.

A while back we switched the user that is used for zopectl commands from
the `zopemaster` user to SpecialUsers.system.

Once the remote requests dispatched from OGDS sync were changed to use the
admin unit auth PAS plugin, they required the userid to generate the
authentication token. This failed, because the SpecialUsers.system does not
have a userid.

This fix therefore explicitly changes the user used for the sync_ogds
zopectl handler to `zopemaster`.

(TTW ogds sync was not affected, because those already were using an actual
user, e.g. an administrator account or `zopemaster`).