Fix encoding issues in classic UI
Since upgrading the Docker base image to Alpine 3.20 (ogcore 2024.15.0
and later) the HTML parser used in plone.app.theming doesn't detect the
correct encoding when parsing a response. As we always use UTF-8, we set
the encoding manually.

Not sure what causes the error, it seems only to happen with newer
Alpine versions.

The behavior can be reproduced with the following snippet, which should
return 'UTF-8', but returns 'ISO-8859-1':

>>> from lxml.html import HTMLParser
>>> parser = HTMLParser()
>>> parser.feed(b'<div><h2>Schlagw\xc3\xb6rter</h2></div>')
>>> parser.close().getroottree().docinfo.encoding
'ISO-8859-1'